Update README.md
Final Project - Breach Response Case Study

SOURCE: Data Breaches That Have Happened in 2022, 2023 and 2024 So Far - https://tech.co/news/data-breaches-updated-list

INSTRUCTIONS:
1. You will begin by identifying a company or affected parties that have incurred a data breach in the last seven years. 
2. Describe the attack category that caused the data breach. 
3. Provide specifics about that attack category including what are the vulnerabilities of that attack category, has there been other examples. 
4. You will give industry statistics that most of them can be found within the X-Force Intelligence report that has been identified as part of this course.
5. Cite sources for your research:  Internet, Wikipedia, Blogs, Journals, Reports, X-Force Threat Intelligence Index 2023, etc.
6. Provide a company description and a breach summary of the incident and data breach. 
7. Make sure to use the case studies that have been presented as part of this course as examples of how to fill out this information that will be reviewed by your peers.

A. TIMELINE
1. What was the series of events leading up to the data breach as well as events that happened after the data breach was identified.  
Provide 6 chronological events or milestones related to this attack.  This can be related events before and after the attack.

Event #1
Event #2
Event #3
Event #4
Event #5
Event #6

2. The data breach identified within the corporation or company? 
3. Was the data breach identified by an outside party? 
4. Include information about when the data breach occurred, when the attack occurred. 
5. What are the summary of the pertinent information around this data breach? 

B. VULNERABILITIES
Provide a summary for 4 specific vulnerabilities for your case in box (4 column table) format.

Vulnerability #1 SUMMARY
Vulnerability #2 SUMMARY
Vulnerability #3 SUMMARY
Vulnerability #4 SUMMARY

1. What was the overall vulnerability for the corporation or company? 
2   Was it potentially that their systems were not hardened? 
-Was it that they were lacking security software? 
-Was it that it took several months to identify the breach? 
3.  You can see that there may be various options for this summary. 
4. Pick out four specific vulnerabilities, anything from people not being educated to, again, not having an anti-virus patch on the server that was attacked. 

Note: You will see that each particular data breach will be unique and you will need to call out specific vulnerabilities for that data breach. 

C. COSTS
2 column table.  1st column lists cost with 1, 2, 3, etc.  Next column to right lists Prevention with 1, 2, 3, etc.

1. Depending on when the data breach occurred, if it's a recent data breach, there may be limited information about the cost to that company or organization.

If the data breach was in the past, you should be able to see a series of different costs that resulted over the years including litigation cost, potential cost to pay out for credit card information that was obtained by the attackers. As well as different costs that the corporation may have incurred in some instances for loss of business. 

D. PREVENTION

In this category, it could be anything from the corporation needed to have an incident response plan.  
Or it could be that the analyst missed several events from the same IP address. 
Again, this is a unique situation and there may be hundreds of different prevention techniques depending on the attack that you select as well as the company data breach that you found in the news.

E. COMPLETION AND PROJECT SUBMISSION

Once you've filled out the template, save or export the presentation to a PDF file and post it back into your course. 
You will next have to peer reviewers take a look at your course and grade you on how much information that you've provided and the accuracy of that information based upon their knowledge.
There will be partial credit for each slide and you will need to obtain 80% on average in order to pass this course.